+++ src/Funds/FundDetails/FundExpenseClasses/index.js
@@ -1,0 +1,1 @@
+export { default as FundExpenseClasses } from './FundExpenseClasses';